responsive layout created

+++ src/Pages/InventoryPage/Inventory/ManageInventory/ManageInventory.jsx
@@ -8,7 +8,7 @@ const ManageInventory = () => {
     const loadData = async () => {
       try {
         const { data } = await axios.get(
-          `http://localhost:5000/inventory-items`
+          `http://localhost:5000/all-inventory-items`
         );
         setEnventoryItems(data);
       } catch (err) {
@@ -18,21 +18,24 @@ const ManageInventory = () => {
     loadData();
   }, []);
   const deleteItem = async (id) => {
-    const url = `http://localhost:5000/inventory-items/${id}`;
-    await axios.delete(url).then((response) => {
-      const { data } = response;
-      if (data) {
-        const remaingProduct = inventoryItemes.filter(
-          (item) => item._id !== id
-        );
-        setEnventoryItems(remaingProduct);
-      }
-    });
+    const sure = window.confirm("Are you sure? You want to delete!");
+    if (sure) {
+      const url = `http://localhost:5000/inventory-items/${id}`;
+      await axios.delete(url).then((response) => {
+        const { data } = response;
+        if (data) {
+          const remaingProduct = inventoryItemes.filter(
+            (item) => item._id !== id
+          );
+          setEnventoryItems(remaingProduct);
+        }
+      });
+    }
   };
   return (
     <>
-      <div className="w-full h-screen bg-gray-100">
-        <div className="max-w-7xl mx-auto sm:px-6 lg:px-8">
+      <div className="w-full min-h-screen px-1 bg-gray-100 my-5 md:my-10">
+        <div className="max-w-7xl hidden md:block mx-auto sm:px-6 lg:px-8">
           <div className="flex flex-col">
             <div className="-mb-2 py-4 flex flex-wrap flex-grow justify-between">
               <div className="flex items-center py-2">
@@ -138,6 +141,75 @@ const ManageInventory = () => {
                             <p>USD</p>
                           </div>
                         </td>
+                        <td className="px-6 py-4 whitespace-no-wrap text-center border-b border-gray-200 text-sm leading-5 font-medium">
+                          <div className="flex justify-center items-center">
+                            <button
+                              onClick={() => deleteItem(item._id)}
+                              className="block  bg-teal-500 hover:bg-teal-600 text-white border-2 border-teal-500 hover:border-teal-600 px-3 py-2 rounded uppercase font-poppins font-medium"
+                            >
+                              Delete
+                            </button>
+                          </div>
+                        </td>
+                      </tr>
+                    ))}
+                  </tbody>
+                </table>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        {/* mobile device  */}
+        <div className="md:hidden w-full  lg:px-8">
+          <div className="flex flex-col">
+            <div className="flex justify-end   items-center py-5">
+              <NavLink
+                to="/dashboard/add-inventory-item"
+                className="inline-block px-4 py-2 border border-transparent text-sm leading-5 font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-500 focus:outline-none focus:shadow-outline"
+              >
+                Create new Items
+              </NavLink>
+            </div>
+
+            <div className="-my-2 py-2 sm:-mx-6 sm:px-6 lg:-mx-8 lg:px-8">
+              <div className="align-middle inline-block w-full shadow overflow-x-auto sm:rounded-lg border-b border-gray-200">
+                <table className="min-w-full table-auto">
+                  <thead>
+                    <tr className="border-b border-gray-200 font-thin bg-white leading-4 tracking-wider text-base text-gray-500">
+                      <th className="px-6 py-5 text-left" colSpan="100%">
+                        <p>Manage Inventory</p>
+                      </th>
+                    </tr>
+                    <tr className="bg-gray-50 border-b border-gray-200 text-xs leading-4 text-gray-500 uppercase tracking-wider">
+                      <th className="px-6 py-3 text-center font-medium">
+                        Product
+                      </th>
+
+                      <th className="px-6 py-3 text-center font-medium">
+                        Update
+                      </th>
+                    </tr>
+                  </thead>
+
+                  <tbody className="bg-white">
+                    {inventoryItemes.map((item) => (
+                      <tr key={item._id}>
+                        <td className="px-6 py-4 whitespace-no-wrap border-b border-gray-200">
+                          <div className="text-sm leading-5 text-gray-900">
+                            <div className="flex flex-col items-center  justify-center">
+                              <div className="flex-shrink-0 h-10 w-10">
+                                <img
+                                  className="h-10 w-10 rounded-full"
+                                  src={item.img}
+                                  alt="product"
+                                />
+                              </div>
+                              <div>{item.name}</div>
+                            </div>
+                          </div>
+                        </td>
+
                         <td className="px-6 py-4 whitespace-no-wrap text-center border-b border-gray-200 text-sm leading-5 font-medium">
                           <div className="flex justify-center items-center">
                             <button
