+++ src/Pages/InventoryPage/MyItems/MyItem.jsx
@@ -31,26 +31,29 @@ const MyItem = () => {
     loadData();
   }, [user]);
   const deleteItem = async (id) => {
-    const url = `http://localhost:5000/inventory-items/${id}`;
-    await axios.delete(url).then((response) => {
-      const { data } = response;
-      if (data) {
-        const remaingProduct = inventoryItemes.filter(
-          (item) => item._id !== id
-        );
-        setEnventoryItems(remaingProduct);
-      }
-    });
+    const sure = window.confirm("Are you sure? You want to delete!");
+    if (sure) {
+      const url = `http://localhost:5000/inventory-items/${id}`;
+      await axios.delete(url).then((response) => {
+        const { data } = response;
+        if (data) {
+          const remaingProduct = inventoryItemes.filter(
+            (item) => item._id !== id
+          );
+          setEnventoryItems(remaingProduct);
+        }
+      });
+    }
   };
   return (
     <>
-      <div className="-my-2 py-2 sm:-mx-6 sm:px-6 lg:-mx-8 lg:px-8">
-        <div className="align-middle inline-block w-full shadow overflow-x-auto sm:rounded-lg border-b border-gray-200">
+      <div className="my-5 max-w-7xl mx-auto  sm:px-6 lg:px-8">
+        <div className="align-middle hidden md:inline-block w-full shadow overflow-x-auto sm:rounded-lg border-b border-gray-200">
           <table className="min-w-full table-auto">
             <thead>
               <tr className="border-b border-gray-200 font-thin bg-white leading-4 tracking-wider text-base text-gray-500">
                 <th className="px-6 py-5 text-left" colSpan="100%">
-                  <p>Manage Inventory</p>
+                  <p className="uppercase">My Items</p>
                 </th>
               </tr>
               <tr className="bg-gray-50 border-b border-gray-200 text-xs leading-4 text-gray-500 uppercase tracking-wider">
@@ -123,6 +126,55 @@ const MyItem = () => {
                     </div>
                   </td>
                   <td className="px-6 py-4 whitespace-no-wrap text-center border-b border-gray-200 text-sm leading-5 font-medium">
+                    <div className="flex justify-center items-center">
+                      <button
+                        onClick={() => deleteItem(item._id)}
+                        className="block  bg-teal-500 hover:bg-teal-600 text-white border-2 border-teal-500 hover:border-teal-600 px-3 py-2 rounded uppercase font-poppins font-medium"
+                      >
+                        Delete
+                      </button>
+                    </div>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
+        {/* mobile screen  */}
+        <div className="align-middle mx-auto inline-block w-full md:hidden shadow bg-white md:bg-transparent  sm:rounded-lg border-b border-gray-200">
+          <table className="w-full table-auto">
+            <thead>
+              <tr className="border-b border-gray-200 font-thin bg-white  tracking-wider text-base text-gray-500">
+                <th className="px-2 py-5 text-left" colSpan="100%">
+                  <p className="uppercase">My Items</p>
+                </th>
+              </tr>
+              <tr className="bg-gray-50 border-b border-gray-200 text-xs text-gray-500 uppercase tracking-wider">
+                <th className="px-2 py-3  text-center font-medium">Product</th>
+
+                <th className="px-2 py-3 text-center font-medium">Update</th>
+              </tr>
+            </thead>
+
+            <tbody className="bg-white">
+              {inventoryItemes.map((item) => (
+                <tr key={item._id}>
+                  <td className="px-6 py-4  border-b border-gray-200">
+                    <div className="text-sm leading-5 text-gray-900">
+                      <div className="flex flex-col items-center  justify-center">
+                        <div className=" h-10 w-10">
+                          <img
+                            className="h-10 w-10 "
+                            src={item.img}
+                            alt="product"
+                          />
+                        </div>
+                        <div>{item.name}</div>
+                      </div>
+                    </div>
+                  </td>
+
+                  <td className=" py-4  text-center border-b border-gray-200 text-sm leading-5 font-medium">
                     <div className="flex justify-center items-center">
                       <button
                         onClick={() => deleteItem(item._id)}
